Extract `FeedStoreSpy` in its own file
